requirements.txt
amqp==2.4.2
apiview==1.3.25
billiard==3.6.0.0
celery==4.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
cryptography==3.2.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==2.2.24
django-celery-beat==1.4.0
django-crontab==0.7.1
django-import-export==2.5.0
django-mysql==3.9.0
django-timezone-field==3.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
jsonfield==3.0.0
kombu==4.5.0
MarkupPy==1.14
mysqlclient==1.3.13
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==2.6.4
pep8==1.7.1
Pillow==7.2.0
pycparser==2.21
PyJWT==1.6.4
python-crontab==2.3.6
python-dateutil==2.8.0
python-decouple==3.1
python-http-client==3.1.0
python3-openid==3.2.0
pytz==2018.7
PyYAML==5.3.1
redis==3.2.1
requests==2.25.1
requests-oauthlib==1.3.1
sendgrid==6.0.4
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.0.2
sqlparse==0.4.1
StringGenerator==0.3.1
tablib==2.0.0
urllib3==1.26.9
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
